fix: moved packages/smart to console
