+++ apps/console/package.json
@@ -49,7 +49,6 @@
     "@spectacular/biome-config": "workspace:*",
     "@spectacular/role-houdini": "workspace:*",
     "@spectacular/skeleton": "workspace:*",
-    "@spectacular/smart": "workspace:*",
     "@spectacular/typescript-config": "workspace:*",
     "@spectacular/utils": "workspace:*",
     "@svelte-plugins/datepicker": "1.0.11",
