+++ apps/console/src/lib/smart/@types/browser-ai.d.ts
