+++ apps/console/src/lib/smart/README.md
