+++ apps/console/src/lib/smart/actions/index.ts
