+++ apps/console/src/lib/smart/actions/smart-paste.ts
