+++ apps/console/src/lib/smart/components/chat.svelte
