+++ apps/console/src/lib/smart/components/constants.test.ts
