+++ apps/console/src/lib/smart/components/constants.ts
