+++ apps/console/src/lib/smart/components/index.ts
