+++ apps/console/src/lib/smart/components/loader-icon.svelte
