+++ apps/console/src/lib/smart/components/magic-spell-textarea.svelte
