+++ apps/console/src/lib/smart/components/result.svelte
