+++ apps/console/src/lib/smart/components/sentiment.svelte
