+++ apps/console/src/lib/smart/components/settings.ts
