+++ apps/console/src/lib/smart/components/smart.ts
