+++ apps/console/src/lib/smart/components/summary-icon.svelte
