+++ apps/console/src/lib/smart/components/support.svelte
