+++ apps/console/src/lib/smart/components/textarea.svelte
