+++ apps/console/src/lib/smart/components/translate-icon.svelte
