+++ apps/console/src/lib/smart/use-object.ts
