+++ apps/console/src/routes/(app)/ai/pastesmart/+page.svelte
@@ -5,8 +5,8 @@ import { getLoadingState } from '$lib/stores/loading';
 import { SlideToggle, getToastStore } from '@skeletonlabs/skeleton';
 import { DebugShell } from '@spectacular/skeleton/components';
 import { Alerts, ErrorMessage } from '@spectacular/skeleton/components/form';
-import { SmartSupport } from '@spectacular/smart';
-import { smartPaste } from '@spectacular/smart/actions';
+import { SmartSupport } from '$lib/smart/components';
+import { smartPaste } from '$lib/smart/actions';
 import { Logger } from '@spectacular/utils';
 import * as Form from 'formsnap';
 import { Loader, MoreHorizontal, Sparkles } from 'lucide-svelte';
