+++ apps/console/src/routes/(app)/ai/writing/+page.svelte
@@ -4,8 +4,8 @@ import { getLoadingState } from '$lib/stores/loading';
 import { getToastStore } from '@skeletonlabs/skeleton';
 import { DebugShell } from '@spectacular/skeleton/components';
 import { Alerts } from '@spectacular/skeleton/components/form';
-import { Support } from '@spectacular/smart';
-import * as Smart from '@spectacular/smart';
+import { Support } from '$lib/smart/components';
+import * as Smart from '$lib/smart/components';
 import { Logger } from '@spectacular/utils';
 import * as Form from 'formsnap';
 import SuperDebug, { defaults, setError, setMessage, superForm } from 'sveltekit-superforms';
