--- packages/smart/biome.jsonc
@@ -1,4 +1,0 @@
-{
-  "$schema": "https://biomejs.dev/schemas/1.9.4/schema.json",
-  "extends": ["@spectacular/biome-config/svelte-library"]
-}
