--- packages/smart/package.json
@@ -1,75 +1,0 @@
-{
-  "name": "@spectacular/smart",
-  "description": "Smart UI components for svelte",
-  "sideEffects": false,
-  "type": "module",
-  "imports": {
-    "#utils": "./src/utils/index.js",
-    "#*": "./src/*"
-  },
-  "exports": {
-    ".": {
-      "types": "./src/components/index.ts",
-      "svelte": "./src/components/index.ts",
-      "default": "./src/components/index.ts"
-    },
-    "./actions": "./src/actions/index.ts",
-    "./actions/*": "./src/actions/*.ts",
-    "./components": "./src/components/index.ts",
-    "./components/*": "./src/components/*/index.ts"
-  },
-  "scripts": {
-    "check": "biome check .",
-    "check:fix": "biome check . --write",
-    "clean": "git clean -xdf",
-    "format": "biome format --write",
-    "lint": "biome lint",
-    "test": "vitest run",
-    "test:unit:coverage": "vitest run --coverage",
-    "test:watch": "vitest"
-  },
-  "devDependencies": {
-    "@ai-sdk/svelte": "1.1.24",
-    "@aibrow/dom-types": "1.3.8",
-    "@floating-ui/dom": "1.6.13",
-    "@inlang/paraglide-sveltekit": "0.16.1",
-    "@skeletonlabs/skeleton": "2.11.0",
-    "@spectacular/biome-config": "workspace:*",
-    "@spectacular/skeleton": "workspace:*",
-    "@spectacular/typescript-config": "workspace:*",
-    "@spectacular/utils": "workspace:*",
-    "@svelte-plugins/datepicker": "1.0.11",
-    "@sveltejs/kit": "2.19.2",
-    "@sveltejs/vite-plugin-svelte": "3.1.2",
-    "@tsconfig/svelte": "5.0.4",
-    "@types/dom-chromium-ai": "0.0.6",
-    "@types/node": "22.13.10",
-    "@vitest/coverage-v8": "3.0.8",
-    "ai": "4.1.61",
-    "chrome-ai": "1.11.1",
-    "date-fns": "4.1.0",
-    "formsnap": "1.0.1",
-    "gjp-4-gpt": "0.0.8",
-    "lucide-svelte": "0.482.0",
-    "svelte": "4.2.19",
-    "svelte-check": "4.1.5",
-    "svelte-select": "5.8.3",
-    "sveltekit-superforms": "2.24.0",
-    "tailwind-merge": "2.6.0",
-    "tailwind-variants": "0.3.1",
-    "tailwindcss": "3.4.17",
-    "tslib": "2.8.1",
-    "typescript": "5.8.2",
-    "vite": "5.4.11",
-    "vitest": "3.0.8",
-    "zod": "3.24.2"
-  },
-  "peerDependencies": {
-    "@sveltejs/kit": "^2.0.0",
-    "svelte": "^4.0.0"
-  },
-  "dependencies": {
-    "@ai-sdk/provider-utils": "2.1.13",
-    "@ai-sdk/ui-utils": "1.1.19"
-  }
-}
