--- packages/smart/tsconfig.json
@@ -1,19 +1,0 @@
-{
-  "extends": "@tsconfig/svelte/tsconfig.json",
-  "include": ["."],
-  "exclude": ["node_modules", "styles", "./*.cjs"],
-  "compilerOptions": {
-    "baseUrl": ".",
-    "allowJs": true,
-    "checkJs": false,
-    "isolatedModules": true,
-    "resolveJsonModule": true,
-    "strict": true,
-    "target": "esnext",
-    "module": "NodeNext",
-    "moduleResolution": "NodeNext",
-    "ignoreDeprecations": "5.0",
-    "lib": ["esnext", "DOM", "DOM.Iterable"],
-    "types": ["vitest/globals", "vitest/importMeta", "@testing-library/jest-dom", "@types/dom-chromium-ai"]
-  }
-}
